v4

+++ src/Http/Api/Controllers/Campaigns/CampaignsController.php
@@ -2,21 +2,25 @@
 
 namespace Spatie\Mailcoach\Http\Api\Controllers\Campaigns;
 
-use Spatie\Mailcoach\Actions\Campaigns\UpdateCampaignAction;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Spatie\Mailcoach\Domain\Campaign\Actions\UpdateCampaignAction;
+use Spatie\Mailcoach\Domain\Campaign\Models\Campaign;
+use Spatie\Mailcoach\Domain\Shared\Traits\UsesMailcoachModels;
 use Spatie\Mailcoach\Http\Api\Controllers\Concerns\RespondsToApiRequests;
 use Spatie\Mailcoach\Http\Api\Requests\CampaignRequest;
 use Spatie\Mailcoach\Http\Api\Resources\CampaignResource;
 use Spatie\Mailcoach\Http\App\Queries\CampaignsQuery;
-use Spatie\Mailcoach\Models\Campaign;
-use Spatie\Mailcoach\Traits\UsesMailcoachModels;
 
 class CampaignsController
 {
-    use UsesMailcoachModels,
+    use AuthorizesRequests,
+        UsesMailcoachModels,
         RespondsToApiRequests;
 
     public function index(CampaignsQuery $campaigns)
     {
+        $this->authorize("viewAny", Campaign::class);
+
         return CampaignResource::collection($campaigns->paginate());
     }
 
@@ -24,6 +28,8 @@ class CampaignsController
         CampaignRequest $request,
         UpdateCampaignAction $updateCampaignAction
     ) {
+        $this->authorize("create", Campaign::class);
+
         $campaignClass = $this->getCampaignClass();
 
         $campaign = new $campaignClass;
@@ -42,6 +48,8 @@ class CampaignsController
         CampaignRequest $request,
         UpdateCampaignAction $updateCampaignAction
     ) {
+        $this->authorize("update", $campaign);
+
         $campaign = $updateCampaignAction->execute(
             $campaign,
             $request->validated(),
@@ -52,6 +60,8 @@ class CampaignsController
 
     public function destroy(Campaign $campaign)
     {
+        $this->authorize("delete", $campaign);
+
         $campaign->delete();
 
         return $this->respondOk();
