+++ src/Http/Api/Controllers/EmailLists/EmailListsController.php
@@ -2,30 +2,39 @@
 
 namespace Spatie\Mailcoach\Http\Api\Controllers\EmailLists;
 
-use Spatie\Mailcoach\Actions\EmailLists\UpdateEmailListAction;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Spatie\Mailcoach\Domain\Audience\Actions\EmailLists\UpdateEmailListAction;
+use Spatie\Mailcoach\Domain\Audience\Models\EmailList;
+use Spatie\Mailcoach\Domain\Shared\Traits\UsesMailcoachModels;
 use Spatie\Mailcoach\Http\Api\Controllers\Concerns\RespondsToApiRequests;
-use Spatie\Mailcoach\Http\Api\Requests\EmailListRequest;
+use Spatie\Mailcoach\Http\Api\Requests\UpdateEmailListSettingsRequest;
 use Spatie\Mailcoach\Http\Api\Resources\EmailListResource;
 use Spatie\Mailcoach\Http\App\Queries\EmailListQuery;
-use Spatie\Mailcoach\Models\EmailList;
-use Spatie\Mailcoach\Traits\UsesMailcoachModels;
 
 class EmailListsController
 {
-    use RespondsToApiRequests, UsesMailcoachModels;
+    use AuthorizesRequests,
+        RespondsToApiRequests,
+        UsesMailcoachModels;
 
     public function index(EmailListQuery $emailLists)
     {
+        $this->authorize("viewAny", EmailList::class);
+
         return EmailListResource::collection($emailLists->paginate());
     }
 
     public function show(EmailList $emailList)
     {
+        $this->authorize("view", $emailList);
+
         return new EmailListResource($emailList);
     }
 
-    public function store(EmailListRequest $request, UpdateEmailListAction $updateEmailListAction)
+    public function store(UpdateEmailListSettingsRequest $request, UpdateEmailListAction $updateEmailListAction)
     {
+        $this->authorize("create", EmailList::class);
+
         $emailListClass = $this->getEmailListClass();
 
         $emailList = new $emailListClass;
@@ -35,8 +44,10 @@ class EmailListsController
         return new EmailListResource($emailList);
     }
 
-    public function update(EmailListRequest $request, EmailList $emailList, UpdateEmailListAction $updateEmailListAction)
+    public function update(UpdateEmailListSettingsRequest $request, EmailList $emailList, UpdateEmailListAction $updateEmailListAction)
     {
+        $this->authorize("update", $emailList);
+
         $emailList = $updateEmailListAction->execute($emailList, $request);
 
         return new EmailListResource($emailList);
@@ -44,6 +55,8 @@ class EmailListsController
 
     public function destroy(EmailList $emailList)
     {
+        $this->authorize("delete", $emailList);
+
         $emailList->delete();
 
         return $this->respondOk();
