+++ src/Http/Api/Controllers/EmailLists/Subscribers/SubscribersController.php
@@ -2,22 +2,27 @@
 
 namespace Spatie\Mailcoach\Http\Api\Controllers\EmailLists\Subscribers;
 
-use Spatie\Mailcoach\Actions\Subscribers\UpdateSubscriberAction;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Spatie\Mailcoach\Domain\Audience\Actions\Subscribers\UpdateSubscriberAction;
+use Spatie\Mailcoach\Domain\Audience\Models\EmailList;
+use Spatie\Mailcoach\Domain\Audience\Models\Subscriber;
+use Spatie\Mailcoach\Domain\Shared\Traits\UsesMailcoachModels;
 use Spatie\Mailcoach\Http\Api\Controllers\Concerns\RespondsToApiRequests;
 use Spatie\Mailcoach\Http\Api\Requests\StoreSubscriberRequest;
 use Spatie\Mailcoach\Http\Api\Resources\SubscriberResource;
 use Spatie\Mailcoach\Http\App\Queries\EmailListSubscribersQuery;
 use Spatie\Mailcoach\Http\App\Requests\EmailLists\Subscribers\UpdateSubscriberRequest;
-use Spatie\Mailcoach\Models\EmailList;
-use Spatie\Mailcoach\Models\Subscriber;
-use Spatie\Mailcoach\Traits\UsesMailcoachModels;
 
 class SubscribersController
 {
-    use UsesMailcoachModels, RespondsToApiRequests;
+    use AuthorizesRequests,
+        UsesMailcoachModels,
+        RespondsToApiRequests;
 
     public function index(EmailList $emailList)
     {
+        $this->authorize("view", $emailList);
+
         $subscribers = new EmailListSubscribersQuery($emailList);
 
         return SubscriberResource::collection($subscribers->paginate());
@@ -25,12 +30,16 @@ class SubscribersController
 
     public function show(Subscriber $subscriber)
     {
+        $this->authorize("view", $subscriber->emailList);
+
         return new SubscriberResource($subscriber);
     }
 
     public function store(StoreSubscriberRequest $request, EmailList $emailList)
     {
-        /** @var \Spatie\Mailcoach\Support\PendingSubscriber $pendingSubscriber */
+        $this->authorize("update", $emailList);
+
+        /** @var \Spatie\Mailcoach\Domain\Audience\Support\PendingSubscriber $pendingSubscriber */
         $pendingSubscriber = $this
             ->getSubscriberClass()::createWithEmail($request->email)
             ->withAttributes($request->subscriberAttributes());
@@ -54,6 +63,8 @@ class SubscribersController
 
     public function destroy(Subscriber $subscriber)
     {
+        $this->authorize("update", $subscriber->emailList);
+
         $subscriber->delete();
 
         return $this->respondOk();
@@ -61,6 +72,8 @@ class SubscribersController
 
     public function update(Subscriber $subscriber, UpdateSubscriberRequest $request, UpdateSubscriberAction $updateSubscriberAction)
     {
+        $this->authorize("update", $subscriber->emailList);
+
         $updateSubscriberAction->execute(
             $subscriber,
             $request->subscriberAttributes(),
