+++ src/Http/App/Controllers/EmailLists/TagsController.php
@@ -2,28 +2,41 @@
 
 namespace Spatie\Mailcoach\Http\App\Controllers\EmailLists;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Spatie\Mailcoach\Domain\Audience\Models\EmailList;
+use Spatie\Mailcoach\Domain\Audience\Models\Tag;
+use Spatie\Mailcoach\Domain\Campaign\Enums\TagType;
 use Spatie\Mailcoach\Http\App\Queries\EmailListTagsQuery;
 use Spatie\Mailcoach\Http\App\Requests\EmailLists\CreateTagRequest;
 use Spatie\Mailcoach\Http\App\Requests\EmailLists\UpdateTagRequest;
-use Spatie\Mailcoach\Models\EmailList;
-use Spatie\Mailcoach\Models\Tag;
 
 class TagsController
 {
+    use AuthorizesRequests;
+
     public function index(EmailList $emailList)
     {
+        $this->authorize('view', $emailList);
+
         $tagsQuery = new EmailListTagsQuery($emailList);
 
-        return view('mailcoach::app.emailLists.tag.index', [
+        return view('mailcoach::app.emailLists.tags.index', [
             'emailList' => $emailList,
             'tags' => $tagsQuery->paginate(),
             'totalTagsCount' => Tag::query()->emailList($emailList)->count(),
+            'totalDefault' => Tag::query()->where('type', TagType::DEFAULT)->emailList($emailList)->count(),
+            'totalMailcoach' => Tag::query()->where('type', TagType::MAILCOACH)->emailList($emailList)->count(),
         ]);
     }
 
     public function store(CreateTagRequest $request, EmailList $emailList)
     {
-        $tag = $emailList->tags()->create(['name' => $request->name]);
+        $this->authorize('update', $emailList);
+
+        $tag = $emailList->tags()->create([
+            'name' => $request->name,
+            'type' => TagType::DEFAULT,
+        ]);
 
         flash()->success(__('Tag :tag was created', ['tag' => $tag->name]));
 
@@ -32,7 +45,9 @@ class TagsController
 
     public function edit(EmailList $emailList, Tag $tag)
     {
-        return view('mailcoach::app.emailLists.tag.edit', [
+        $this->authorize('update', $emailList);
+
+        return view('mailcoach::app.emailLists.tags.edit', [
             'emailList' => $emailList,
             'tag' => $tag,
         ]);
@@ -40,6 +55,8 @@ class TagsController
 
     public function update(UpdateTagRequest $request, EmailList $emailList, Tag $tag)
     {
+        $this->authorize('update', $emailList);
+
         $tag->update([
             'name' => $request->name,
         ]);
@@ -51,6 +68,8 @@ class TagsController
 
     public function destroy(EmailList $emailList, Tag $tag)
     {
+        $this->authorize('update', $emailList);
+
         $tag->delete();
 
         flash()->success(__('Tag :tag was deleted', ['tag' => $tag->name]));
